stopping work on this branch
There's a fundamental bug in IE8:
element.innerHTML does not preserve comments.

jtmpl relies on this behavior to parse a template string.

It does not seem IE8 exposes HTML parsing JavaScript API.

This still can be fixed by leveraging a JavaScript parser,
but this is too much to bother. IE8, rest in pieces.